__repr__ returns an "un-repr-able" bytestring latin1 chars in filename, #36
An initial attempt at a fix passed all tests, but the origin issue
still failed in REPL.  Punting by returning a bytestring that cannot be
repr-ed back to an object if the filename has latin1 characters that
the ASCII code cannot encode.  All other cases working as expected.
Still needs some thought as to the proper test.