Update Cygwin, macports issues in doc
The macports issue was fixed upstream by the exempi port maintainer.

Make it clear that the Mac OS X installation instructions are not
meant for Homebrew or MacPorts users.

Cygwin mostly works as a platform, had been under the impression that
it did not.  Change the documentation to reflect what works and what
does not.

"files_check_file_format" and "files_get_file_info" are the two exempi
library functions that do not seem to be present in the Cygwin version
of the library.

Skip the one test that absolutely cannot run on Cygwin.  Two others can
run certain portions on Cygwin, so they are not completely filtered out.

Some tests were checking the file format of BlueSquare.jpg, which will
fail on Cygwin.  Since that specific test is done in
test.test_exempi.TestSuite.test_formats, those checks can be removed.

Two tests added to nail down Cygwin behavior.

Closes-Issue: #6
Closes-Issue: #41